[By-257] Move: auto_information 디렉토리로 변경
