+++ lib/widget/auto_information/autoInfo_image.dart
